+++ engine/src/main/kotlin/engine/patterns/BasicPatterns.kt
@@ -42,7 +42,7 @@ class ConstantPattern : BasePattern() {
 class VariableExpressionPattern : BasePattern() {
     override fun doFindMatches(context: Context, match: Match, subexpression: Expression): Sequence<Match> {
         return when {
-            !subexpression.isConstantIn(context.solutionVariables) -> sequenceOf(match.newChild(this, subexpression))
+            !subexpression.isConstant() -> sequenceOf(match.newChild(this, subexpression))
             else -> emptySequence()
         }
     }